Remove redundant code and merge rest into rt-std
