--- pwasm-libc/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "pwasm-libc"
-version = "0.1.0"
-authors = ["Sergey Pepyakin <s.pepyakin@gmail.com>"]
-license = "MIT/Apache-2.0"
-readme = "README.md"
-repository = "https://github.com/paritytech/pwasm-std"
-homepage = "https://github.com/paritytech/pwasm-std"
-documentation = "https://paritytech.github.io/pwasm-std/pwasm_std/"
-description = "Parity WebAssembly standard library libc bindings"
-keywords = ["wasm", "parity", "webassembly", "blockchain"]
-categories = ["no-std", "embedded"]
-
-[features]
-strict = []
